requirements.txt
asgiref==3.6.0
asyncio==3.4.3
celery==5.2.7
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
httpx==0.23.3
hupper==1.12
nltk==3.8.1
python-dotenv==1.0.0
uvicorn==0.21.1
flower==1.0.0
redis==3.5.3
aredis==1.1.8
aioboto3==11.1.0
adrf==0.1.1
psycopg[pool,c]==3.1.8
django-allauth==0.54.0
gunicorn==20.1.0
django-async-orm==0.1.11
